renames useEgoToken to useAuthContext
